tests(dashboard): add integration tests
